+++ cores/arduino/USBAPI.h
@@ -59,7 +59,7 @@ class CDC : public HardwareSerial
 public:
     CDC(struct _stm32l0_usbd_cdc_t *usbd_cdc, void (*serialEventRun)(void));
     void begin(unsigned long baudRate);
-    void begin(unsigned long baudrate, uint16_t config);
+    void begin(unsigned long baudrate, uint32_t config);
     void end(void);
 
     virtual int available(void);
